requirements.txt
django==3.0.10
requests==2.24.0
djangorestframework==3.12.1
Pillow==7.2.0